chore: remove dead services, tools and stray files
tools/_legacy held 13 modules. Twelve were referenced by nothing at all.
The thirteenth was live: tools/splunk.py symlinked to _legacy/splunk.py
and mcp-config.json launches that path, so promote it to a real file at
tools/splunk.py and drop _legacy entirely.

Its test asserted the symlink — the mechanism — rather than the
guarantee mcp-config.json actually depends on, so it now checks that the
configured path resolves.

cribl_service.py and investigation_workflow_service.py had no importers,
and the two root files were one-off scratch output.
